// src/app/modules/clients/model/CareerOrWorkModel.ts
import { MlcModel } from './common/mlc.model';

export class CareerOrWork implements MlcModel {
  name: string;
  jobtitle: string;
  jobdescription: string;
  locationoraddress: string;
  interviewdate: string;
  interviewernames: string;
  dateofoffer: string;
  dateofacceptance: string;
  employeeid: string;
  employeeaccountorlogin: string;
  datestarted: string;
  datefinishedorexpectedfinish: string;
  highlights: string;
  learnings: string;
  experiences: string;
  coursesandcertifications: string;
  awards: string;
  parkingdetails: string;
  accommodationname: string;
  accommodationaddress: string;
  traveldetails: string;
  notes: string;
}